backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
httpx
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
sqlalchemy
python-dateutil
